[IMP] rest_log: handle odoo.http.Response objects in logs
Some endpoints will not return a dict as expected by the ``_dispatch_with_db_logging`` method (and the other methods called within), but an object of type ``odoo.http.Response``.
When that happens, the instance fails to handle it while creating logs (ie: tries to call ``json.dumps()`` on such object, which is not allowed).
This patch should fix the issue.